refactor(serial): Move serial back into main crate
Move the serial back into the main crate.
