+++ src/serial/registers/ier.rs
@@ -1,6 +1,6 @@
 //! A module containing the operations, internal fields and flags accessible for
 //! a [`InterruptEnableRegister`].
-use crate::ComPort;
+use crate::serial::ComPort;
 use arch::io::{
     port::Port,
     register::{ReadRegister, Register, WriteRegister},
