+++ src/serial/registers/iir.rs
@@ -1,6 +1,6 @@
 //! A module containing the operations, internal fields and flags aaccessible
 //! for an [`InterruptIdentificationRegister`].
-use crate::ComPort;
+use crate::serial::ComPort;
 use arch::io::{
     port::Port,
     register::{ReadRegister, Register},
